+++ CHANGELOG.md
@@ -15,6 +15,11 @@ You can find our backwards-compatibility policy [here](https://github.com/hynek/
 
 ## [Unreleased](https://github.com/hynek/stamina/compare/25.1.0...HEAD)
 
+### Changed
+
+- Raise `ValueError` when *attempts* and *timeout* are both `None` to prevent unbounded stop conditions.
+  [#101](https://github.com/hynek/stamina/pull/101)
+
 
 ## [25.1.0](https://github.com/hynek/stamina/compare/24.3.0...25.1.0) - 2025-03-12